HLRA-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
pytesseract
PyPDF2
Pillow
python-dotenv
pytest
httpx
pdf2image
email-validator
aiofiles
pdfplumber
motor
pymongo
pytest
pytest-asyncio
pytest-cov
httpx
bcrypt